requirements.txt
fastapi
uvicorn
github-copilot-sdk
python-multipart
streamlit
requests